Task 4.21-4.22 Updated delete operation to be performed only by authorized user

+++ part4/blog_backend/controllers/blogs.js
@@ -1,7 +1,5 @@
 const blogRouter = require('express').Router()
-const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
-const User = require('../models/user')
 
 blogRouter.get('/', async (_request, response) => {
   const blogs = await Blog
@@ -11,29 +9,21 @@ blogRouter.get('/', async (_request, response) => {
 })
 
 blogRouter.post('/', async (request, response) => {
-  const {body, token} = request
+  const {body, user} = request
 
-  // checking of the validity of the token
-  // decoding the token and returning the Object which the token was based on
-  const decodedToken = jwt.verify(token, process.env.SECRET)
-  const {id: userId} = decodedToken
-
-  if (!userId) {
+  if (!user) {
     return response.status(401).json({
       error: 'token missing or invalid'
     })
   }
 
-  const user = await User.findById(userId)
-
   const blog = new Blog({
     ...body,
     likes: body.likes ? body.likes : 0,
-    user: user._id
+    user: user.id
   })
 
   const savedBlog = await blog.save()
-
   user.blogs = user.blogs.concat(savedBlog._id)
   await user.save()
 
@@ -41,19 +31,15 @@ blogRouter.post('/', async (request, response) => {
 
 })
 
-// // error handling using next func without express-async-errors lib
-// blogRouter.post('/', async (request, response, next) => {
-//   const blog = new Blog(request.body)
-//   try {
-//     const result = await blog.save()
-//     response.status(201).json(result)
-//   } catch (error) {
-//     next(error)
-//   }
-// })
-
+// error handling using next func without express-async-errors lib
 blogRouter.get('/:id', async (request, response, next) => {
-  const {id} = request.params
+  const {params: {id}, user} = request
+  if (!user) {
+    return response.status(401).json({
+      error: 'token missing or invalid'
+    })
+  }
+
   try {
     const returnedBlog = await Blog.findById(id)
     if (returnedBlog) {
@@ -68,8 +54,26 @@ blogRouter.get('/:id', async (request, response, next) => {
 
 blogRouter.put('/:id', async (request, response, next) => {
   const {id} = request.params
-  const {body} = request
+  const {body, user} = request
+  if (!user) {
+    return response.status(401).json({
+      error: 'token missing or invalid'
+    })
+  }
+
   try {
+    const blogToUpdate = await Blog.findById(id)
+
+    if (!blogToUpdate) {
+      return response.status(404).end()
+    }
+
+    if (blogToUpdate.user?.toString() !== user.id) {
+      return response.status(401).json({
+        error: 'only the creator can delete a blog'
+      })
+    }
+
     const updatedBlog = await Blog.findByIdAndUpdate(
       id,
       {...body},
@@ -81,21 +85,46 @@ blogRouter.put('/:id', async (request, response, next) => {
     } else {
       response.json(updatedBlog)
     }
+
   } catch (e) {
     next(e)
   }
 })
 
-
 blogRouter.delete('/:id', async (request, response, next) => {
-  const {id} = request.params
+  const {params: {id}, user} = request
+  if (!user) {
+    return response.status(401).json({
+      error: 'token missing or invalid'
+    })
+  }
+
   try {
-    const returnedData = await Blog.findByIdAndRemove(id)
-    if (returnedData === null) {
+    const blogToDelete = await Blog.findById(id)
+
+    if (!blogToDelete) {
+      return response.status(404).end()
+    }
+
+    // id is a Mongoose Object and needs to be convert to string
+    if (blogToDelete.user?.toString() !== user.id) {
+      return response.status(401).json({
+        error: 'only the creator can delete a blog'
+      })
+    }
+
+    const result = await Blog.findByIdAndRemove(id)
+
+    if (result === null) {
       response.status(404).end()
     } else {
       response.status(204).end()
     }
+
+    // blog id is a Mongoose Object not a String and needs to be convert to string
+    user.blogs = user.blogs.filter(blogId => blogId.toString() !== result._id.toString())
+
+    await user.save()
   } catch (e) {
     next(e)
   }
